Moved test cases of frame skipping pp to another file
